[refactor] 로그인, 회원가입 페이지 약간의 수정
