+++ src/app/router/AppRouter.tsx
@@ -2,8 +2,8 @@ import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 
 import HomePage from '@/page/home/HomePage';
 
-import Login from '@/page/auth/Login';
-import SignUp from '@/page/auth/SignUp';
+import LoginPage from '@/page/auth/ui/LoginPage';
+import SignUpPage from '@/page/auth/ui/SignUpPage';
 import MakeChallengePage from '@/page/makeChallenge/MakeChallengePage';
 import JoinChallengePage from '@/page/joinChallenge/JoinChallengePage';
 
@@ -18,11 +18,11 @@ export const AppRouter = () => {
         { path: '', element: <HomePage /> },
         {
           path: 'login',
-          element: <Login />,
+          element: <LoginPage />,
         },
         {
           path: 'signup',
-          element: <SignUp />,
+          element: <SignUpPage />,
         },
         {
           path: 'make-new-challenge',
