+++ src/page/auth/utils/regex.ts
@@ -1,0 +1,2 @@
+export const nicknameRegex = /^[a-zA-Z0-9가-힣]{2,10}$/;
+export const passwordRegex = /^(?=.*[A-Za-z])(?=.*\d)(?=.*[!@#$%^&*?])[A-Za-z\d!@#$%^&*?]{8,20}$/;
